用户问题修改，创建用户之后userId返回问题 默认地址问题修改，前端的用户校验和接口请求为并行，非常规的访问方式访问下单页存在cookie不存在的情况，处理方式是直接返回代码1022，重新登录

+++ src/main/java/com/quxin/freshfun/service/user/UserBaseService.java
@@ -14,7 +14,7 @@ public interface UserBaseService {
      * @param userBase  用户信息对象
      * @return 受影响行数
      */
-    Integer addUserBaseInfo(UserBasePOJO userBase);
+    Long addUserBaseInfo(UserBasePOJO userBase);
 
     /**
      * 为用户添加父级捕手id
